Link against __real_memcpy when not using glibc.
The memcpy backwards compatibility wrapper for older versions of glibc links
against a specific symbol that is not present in other libc implementations
(e.g. musl libc).

This commit changes the wrapper to only be defined for glibc.
`-Wl,-wrap,memcpy` should not be passed as it is not needed for other libc
implementations.